Enabling MyLocationOverlay on MapView
MyLocationOverlay was not displaying on the MapView. It turns out
it needs to be added to the MapView's layers. Additionally, the
MyLocationOverlay was complaining to the log about requiring
ACCESS_FINE_LOCATION permissions.

MyLocationOverlay is now visible on the MapView in the emulator,
if you use DDMS to send a location to the emulator.